fix xaxis style, move progress computing to tasks
