+++ src/views/ReportsView.vue
@@ -19,6 +19,10 @@ const heatmapOptions = {
       hideOverlappingLabels: true,
       // format: 'dd.MM',
       formatter: (value, timestamp) => date.formatDate(timestamp, 'DD.MM'),
+      style: {
+        colors: 'rgba(0, 0, 0, 0.54)',
+        fontSize: '.75rem'
+      },
       offsetY: -4,
       offsetX: 5
     },
